refactor: extract overlays/shared/ElementNode
