+++ packages/diagram/src/overlays/edge-details/EdgeDetailsXYFlow.tsx
@@ -16,7 +16,7 @@ import { SelectEdge } from './SelectEdge'
 import * as css from './SelectEdge.css'
 import { useLayoutedEdgeDetails, ZIndexes } from './use-layouted-edge-details'
 import { CompoundNode } from '../shared/xyflow/CompoundNode'
-import { ElementNode } from './xyflow/ElementNode'
+import { ElementNode } from '../shared/xyflow/ElementNode'
 import { RelationshipEdge } from './xyflow/RelationshipEdge'
 import type { BaseTypes } from '../shared/_types'
 import { only } from 'remeda'
