+++ packages/diagram/src/overlays/relationships-of/RelationshipsXYFlow.tsx
@@ -22,7 +22,7 @@ import { cssReactflowMarker } from '../Overlays.css'
 import { type RelationshipsOfTypes } from './_types'
 import { ZIndexes } from './use-layouted-relationships'
 import { CompoundNode } from '../shared/xyflow/CompoundNode'
-import { ElementNode } from './xyflow/ElementNode'
+import { ElementNode } from '../shared/xyflow/ElementNode'
 import { EmptyNode } from './xyflow/EmptyNode'
 import { RelationshipEdge } from './xyflow/RelationshipEdge'
 import { type BaseTypes } from '../shared/_types'
